Lower throws call tree analysis depth to 0
By default, recent versions of PHPStorm set the depth for throws call tree analysis to 1. This conflicts with the coding standards provided by `mediact/testing-suite`. A call tree depth of 1 or greater makes it so exceptions thrown by other code called within the current method must be either caught or documented using an `@throws` annotation in the method docblock. The MediaCT coding standards only require exceptions thrown in the current level of the stack to be documented.